Fixed some matrix displaying
matrix is much better at displaying significant values

+++ OpenGLearning/OpenGLearning/Camera.h
@@ -113,7 +113,7 @@ public:
 	{
 		this->position = NewPos;
 	}
-	void updateMouseInput(const float& dt, const double& offsetX, const double& offsetY)
+	void updateMouseInput(const float& dt, const float& offsetX, const float& offsetY)
 	{
 		//Update pitch, yaw and roll
 		this->pitch += static_cast<GLfloat>(offsetY) * this->sensitivity *dt;
